ДЗ №8: +++ (сделал на прошлом или позапрошлом дз) 1. Сделайте страницу для отображения всех товаров. +++ (добавил, но ест вопрос - ниже) 2. Добавьте форму с фильтром по минимальной и/или максимальной цене. --- 3. * Попробуйте реализовать разбивку всех товаров на страницы, по 10 товаров на каждой.
??? Почему-то слетело отображение на фронте price (BigDecimal). Они не приходят в List от бэка, хотя в базе присутствуют. Пока причину сбоя не нашел

+++ src/main/java/edu/geekbrains/repositories/ProductRepository.java
@@ -4,11 +4,12 @@ import edu.geekbrains.entities.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    List<Product> findAllByPriceBetween(Integer min, Integer max);
+    List<Product> findAllByPriceBetween(BigDecimal min, BigDecimal max);
 
 }
